feat: support torchscript extension
